Use iid-verify directly and check for correct region
