+++ lib/api.js
@@ -2,12 +2,34 @@
 const APIBuilder = require('taskcluster-lib-api');
 const assert = require('assert');
 const crypto = require('crypto');
-const {validate} = require('./identity-document');
 const taskcluster = require('taskcluster-client');
+const verify = require('iid-verify');
 
 const {getQueueStats, getQueueUrl, purgeQueue} = require('sqs-simple');
 
 const log = require('./log');
+
+const AWS_PUBLIC_CERTIFICATE = {
+  'us-east-1':
+`-----BEGIN CERTIFICATE-----
+MIIC7TCCAq0CCQCWukjZ5V4aZzAJBgcqhkjOOAQDMFwxCzAJBgNVBAYTAlVTMRkw
+FwYDVQQIExBXYXNoaW5ndG9uIFN0YXRlMRAwDgYDVQQHEwdTZWF0dGxlMSAwHgYD
+VQQKExdBbWF6b24gV2ViIFNlcnZpY2VzIExMQzAeFw0xMjAxMDUxMjU2MTJaFw0z
+ODAxMDUxMjU2MTJaMFwxCzAJBgNVBAYTAlVTMRkwFwYDVQQIExBXYXNoaW5ndG9u
+IFN0YXRlMRAwDgYDVQQHEwdTZWF0dGxlMSAwHgYDVQQKExdBbWF6b24gV2ViIFNl
+cnZpY2VzIExMQzCCAbcwggEsBgcqhkjOOAQBMIIBHwKBgQCjkvcS2bb1VQ4yt/5e
+ih5OO6kK/n1Lzllr7D8ZwtQP8fOEpp5E2ng+D6Ud1Z1gYipr58Kj3nssSNpI6bX3
+VyIQzK7wLclnd/YozqNNmgIyZecN7EglK9ITHJLP+x8FtUpt3QbyYXJdmVMegN6P
+hviYt5JH/nYl4hh3Pa1HJdskgQIVALVJ3ER11+Ko4tP6nwvHwh6+ERYRAoGBAI1j
+k+tkqMVHuAFcvAGKocTgsjJem6/5qomzJuKDmbJNu9Qxw3rAotXau8Qe+MBcJl/U
+hhy1KHVpCGl9fueQ2s6IL0CaO/buycU1CiYQk40KNHCcHfNiZbdlx1E9rpUp7bnF
+lRa2v1ntMX3caRVDdbtPEWmdxSCYsYFDk4mZrOLBA4GEAAKBgEbmeve5f8LIE/Gf
+MNmP9CM5eovQOGx5ho8WqD+aTebs+k2tn92BBPqeZqpWRa5P/+jrdKml1qx4llHW
+MXrs3IgIb6+hUIB+S8dz8/mmO0bpr76RoZVCXYab2CZedFut7qc3WUH9+EUAH5mw
+vSeDCOUMYQR7R9LINYwouHIziqQYMAkGByqGSM44BAMDLwAwLAIUWXBlk40xTwSw
+7HX32MxXYruse9ACFBNGmdX2ZBrVNGrN9N2f6ROk0k9K
+-----END CERTIFICATE-----`,
+};
 
 let builder = new APIBuilder({
   title: 'EC2 Instance Manager',
@@ -33,6 +55,7 @@ let builder = new APIBuilder({
   ],
   errorCodes: {
     AuthorizationFailed: 403,
+    CertificateNotFound: 404,
   },
 });
 
@@ -561,7 +584,18 @@ api.declare({
   ].join(' '),
 }, async function(req, res) {
   let strDoc = Buffer.from(req.body.document, 'base64').toString();
-  if (!validate(strDoc, req.body.signature)) {
+  let doc = JSON.parse(strDoc);
+  let cert = AWS_PUBLIC_CERTIFICATE[doc.region];
+
+  if (!cert) {
+    log.error({
+      signature: req.body.signature,
+      document: strDoc,
+    }, 'The certificate for this region was not found');
+    return res.reportError('CertificateNotFound', 'Failure to verify authorization');
+  }
+
+  if (!verify(cert, strDoc, req.body.signature)) {
     log.error({
       signature: req.body.signature,
       document: strDoc,
@@ -569,7 +603,6 @@ api.declare({
     return res.reportError('AuthorizationFailed', 'Failure to verify authorization');
   }
 
-  let doc = JSON.parse(strDoc);
   this.state.claimCredentials({
     region: doc.region,
     id: doc.instanceId,
@@ -589,16 +622,24 @@ api.declare({
   ].join(' '),
 }, async function(req, res) {
   let strDoc = Buffer.from(req.body.document, 'base64').toString();
+  let doc = JSON.parse(strDoc);
+  let cert = AWS_PUBLIC_CERTIFICATE[doc.region];
 
-  if (!validate(strDoc, req.body.signature)) {
+  if (!cert) {
+    log.error({
+      signature: req.body.signature,
+      document: strDoc,
+    }, 'The certificate for this region was not found');
+    return res.reportError('CertificateNotFound', 'Failure to verify authorization');
+  }
+
+  if (!verify(cert, strDoc, req.body.signature)) {
     log.error({
       signature: req.body.signature,
       document: strDoc,
     }, 'Failure to verify signature');
     return res.reportError('AuthorizationFailed', 'Failure to verify authorization');
   }
-
-  let doc = JSON.parse(strDoc);
 
   if (!await this.state.canClaimCredentials({region: doc.region, id: doc.instanceId})) {
     return res.reportError('AuthorizationFailed', 'Failure to verify authorization');
